+++ services/server/src/server/services/utils/contract-creation-util.ts
@@ -1,6 +1,7 @@
 import type { SourcifyChain } from "@ethereum-sourcify/lib-sourcify";
 import { StatusCodes } from "http-status-codes";
 import logger from "../../../common/logger";
+import { deriveEtherscanApiKey } from "./etherscan-util";
 
 const ETHERSCAN_API =
   "https://api.etherscan.io/v2/api?chainid=${CHAIN_ID}&module=contract&action=getcontractcreation&contractaddresses=${ADDRESS}&apikey=";
@@ -271,10 +272,7 @@ export const getCreatorTx = async (
     sourcifyChain.fetchContractCreationTxUsing?.etherscanApi &&
     sourcifyChain?.etherscanApi?.supported
   ) {
-    const apiKey =
-      process.env[sourcifyChain.etherscanApi.apiKeyEnvName || ""] ||
-      process.env.ETHERSCAN_API_KEY ||
-      "";
+    const apiKey = deriveEtherscanApiKey(sourcifyChain);
     const fetcher = getEtherscanApiContractCreatorFetcher(
       apiKey,
       sourcifyChain.chainId,